requirements.txt
langchain
langchain-ollama
langchain-qdrant
gitingest
python-dotenv
qdrant-client